Completed algorithm for deducing plan steps for requested parameter values in selected tasks
Yay! This finally works! You can ask for an experience-based solution on how to solve a task, based on its pattern (e.g. "fetch-and-place") and the parameters you're interested in ({'object': 'cheese', 'destination': 'table'}). No scoring for now, though; they are not yet in the source data, but will soon be. So the ordering of the solutions is rather random at the moment.