Using multiple scoring methods (default: `OSPT+LPFO`); fixed bug when no bindings
A bug that occurred when a function has no bindings at all (because it has no parameters) would always result in the plan being discarded: `"".split(" ")` returns `['']` instead of `[]`; '' is the prologified to '?' and counts as parameter that should be bound in the configuration parameter space. It naturally is not, so the plan is discarded as invalid. This is fixed now by checking for an empty initial string.